+++ local_image_provider_platform_interface/lib/local_image_provider_platform_interface.dart
@@ -104,4 +104,8 @@ abstract class LocalImageProviderPlatform extends PlatformInterface {
   Future cleanup() {
     throw UnimplementedError();
   }
+
+  Future<String> newAlbum(String title, bool shared) {
+    throw UnimplementedError();
+  }
 }