+++ local_image_provider_platform_interface/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.1.0
+
+* new interface for album creation
+* new interface for adding pictures to albums
+
 ## 2.0.1
 
 * fix for null safe handling of missing video file